requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain-community
langchain-google-genai
faiss-cpu
